fix(quick-260704-0v6): first-compile fixes from CI run 28688370751
- profile.rs imported JobQueue from nameless_core::ports; it lives in
  nameless_core::job (E0432, broke the lean build)
- queue_sqlxmq.rs cast retries to usize; sqlxmq 0.6 set_retries takes
  u32 (E0308)
- nameless-adapters used uuid::Uuid in repo_pg.rs without declaring the
  dep (E0433) - now optional behind the postgres feature like its peers

+++ crates/nameless-adapters/src/queue_sqlxmq.rs
@@ -133,8 +133,8 @@ impl JobQueue for SqlxmqJobQueue {
         let payload =
             serde_json::to_value(&env).map_err(|e| JobError::Serialization(e.to_string()))?;
         let backoff = self.policy.backoff(0);
-        // Mirror the RetryPolicy onto sqlxmq's own retry machinery.
-        let retries = self.policy.max_attempts.saturating_sub(1) as usize;
+        // Mirror the RetryPolicy onto sqlxmq's own retry machinery (set_retries takes u32).
+        let retries = self.policy.max_attempts.saturating_sub(1);
 
         // Select the per-kind builder (distinct job name + channel). The builder type is the same
         // (`sqlxmq::JobBuilder`) across all registrations — only the embedded name/channel differ —
